% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_distribution.R, R/stat_distribution.R
\docType{data}
\name{GeomDistribution}
\alias{GeomDistribution}
\alias{geom_distribution}
\alias{StatDistribution}
\alias{stat_distribution}
\title{Draw a probability density function as a continuous curve}
\usage{
geom_distribution(
  mapping = NULL,
  data = NULL,
  stat = "distribution",
  position = "identity",
  ...,
  fill = NA,
  outline.type = "upper",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_distribution(
  mapping = NULL,
  data = NULL,
  geom = "ribbon",
  position = "identity",
  ...,
  fun,
  xlim = NULL,
  n = 500,
  p_limit = 0.001,
  location = 0,
  fill = NA,
  alpha = 0.3,
  outline.type = "upper",
  color = "black",
  size = 0.5,
  args = list(),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{Ignored by \code{stat_distribution()}, do not use.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{fill}{Color of fill under the curve. No fill (\code{NA}) by default.}

\item{outline.type}{Type of the outline of the area; \code{"both"} draws both the
upper and lower lines, \code{"upper"}/\code{"lower"} draws the respective lines only.
\code{"full"} draws a closed polygon around the area.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{geom}{The geometric object to use display the data}

\item{fun}{Function to use. A quoted or character name referencing a probability
density function; see examples. Must be vectorised.}

\item{xlim}{Optionally, restrict the range of the function to this range.}

\item{n}{Number of points to interpolate along the main body of the
probability density.}

\item{p_limit}{The lower probability density limit to plot.}

\item{location}{Where to shift the origin. For probability density functions
that do not inherently allow for changing their location, this allows that
shift.}

\item{alpha}{Fill color transparency, if applicable.}

\item{color}{Color of the line.}

\item{size}{Size of the line.}

\item{args}{List of additional arguments passed on to the function defined by \code{fun}.}
}
\description{
Computes and draws a probability density function as a continuous curve.
}
\section{Aesthetics}{

\code{geom_distribution()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x} \emph{or} \code{y}}
\item \strong{\code{ymin} \emph{or} \code{xmin}}
\item \strong{\code{ymax} \emph{or} \code{xmax}}
\item \code{alpha}
\item \code{colour}
\item \code{fill}
\item \code{group}
\item \code{linetype}
\item \code{size}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\section{Computed variables}{

\code{stat_distribution()} computes the following variables:
\describe{
\item{x}{x values along a grid where the probability density is at least \code{p_limit}}
\item{y}{value of the probability density function evaluated at corresponding x}
}
}

\examples{

# geom_distribution() is useful for overlaying probability distributions
set.seed(1234)
ggplot(data.frame(x = rnorm(100)), aes(x)) +
  geom_density() +
  geom_distribution(fun = dnorm, color = "red")

}
\keyword{datasets}
