% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pdensity.R
\name{plot_pdensity}
\alias{plot_pdensity}
\title{Quickly plot probability density function}
\usage{
plot_pdensity(
  fun,
  x = NULL,
  color = "skyblue",
  linetype = "solid",
  size = 1.5,
  fill = NA,
  alpha = 0.3,
  location = 0,
  p_limit = 0.001,
  lower_bound = -10000,
  upper_bound = 10000,
  ...
)
}
\arguments{
\item{fun}{A probability density function (e.g. \code{dnorm}).}

\item{x}{(Optional) A sequence of values for the probability density
function. If not supplied, the function will be plotted for all points
with probability density greater than \code{p_limit} (default 0.001)
between the \code{lower_bound} (default -10000) and the \code{upper_bound}
(default 10000).}

\item{color}{The color of the plotted line (default "skyblue").}

\item{linetype}{The linetype (default "solid").}

\item{size}{The size of the line (default 1.5).}

\item{fill}{The color of the fill under the line (default NA = no fill)}

\item{alpha}{The alpha transparency for the fill (default 0.3).}

\item{location}{For probability density functions that do not have a
location parameter, this will shift the center (zero) of that distribution
to the specified location.}

\item{p_limit}{The probability density cutoff for plotting.}

\item{lower_bound}{The (potential) lower range of plotting subject to the
\code{p_limit} cutoff (default -10000).}

\item{upper_bound}{The (potential) upper range of plotting subject to the
\code{p_limit} cutoff (default 10000).}

\item{...}{Any additional parameters will be passed to the probability
density function}
}
\value{
A ggplot \code{gg} object.
}
\description{
Quickly plot probability density function
}
\examples{
plot_pdensity(dnorm)
plot_pdensity(dnorm, seq(-3, 3, length.out = 100))
plot_pdensity(dbeta, shape1 = 2, shape2 = 2)
}
